+++ mobile/app/src/main/java/com/lacuc/pets/data/login/LoginService.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.data
+package com.lacuc.pets.data.login
 
 import retrofit2.http.*
