+++ mobile/app/src/main/java/com/lacuc/pets/data/post/PostService.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.data
+package com.lacuc.pets.data.post
 
 import retrofit2.http.*
